Few changes and additions
I've modified AppointmentRequestWidget to inherit from RequestWidgetBase
instead of QWidget. It's better that way, because it's realy a request
widget. The AppointmentWidget ui was modified accordingly. I've started
to implement some methods on the AppointmentRequestWidget but realized,
that I'm going to need a mechanism to answer the questions from a service
for that. Dinner time.

 On branch master
 Changes to be committed:
	modified:   appointment_and_request_creator.pro
	modified:   appointment_request_widget.cc
	modified:   appointment_request_widget.h
	modified:   appointment_request_widget.ui
	modified:   appointment_widget.cc
	modified:   appointment_widget.h
	modified:   appointment_widget.ui
	modified:   request_widget_base.h