The appointment request creation works, and can be saved.
Note: the dates and times of the appointment request shouldn't
be part of the service configurator. They aren't that closely
realted, but there is no data structure to represent a configured
service without the dates. I could use the AppointmentRequest
class, with dates and times with no meaning, and create
another fully filled AppointmenRequest later. Will see.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   question_display_widget.cc
	modified:   service_configurator_widget.cc