I've made some changes to the service configurator. Detes below.
The service configurator has to be called from the AppointmentRequestWidget,
so I've changed the on_actionNew_AppointmentRequest_trigger() to change the
request_widget. Further changes have to be made however, because that way
way the API of the ServiceConfiguratorWidget would be more similar to the
BaseRequestWidget API, and that way their functionality could be more
in sync too.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   appointment_request_widget.cc
	modified:   appointment_request_widget.h
	modified:   appointment_request_widget.ui
	modified:   main_window.cc
	modified:   main_window.h
	modified:   main_window.ui
	modified:   service_configurator_widget.cc
	modified:   service_configurator_widget.h
	modified:   service_configurator_widget.ui